requirements.txt
feedparser
newspaper3k
Pillow
requests
langchain
langchain-groq
python-dotenv
lxml_html_clean
gradio
torch